New overall reaction calculation for FXS (dirty/back of the envelope)
Uses as many trials for each rat based on all Base cases with the same number of days for each frequency (so minimum day count for either 4, 8, 16, or 32)

Uses all base cases which may not be ideal since some are after re-establishing baseline (i.e. following Catch) which can change the slope.